Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main (){
-    int r, area;
-    scanf("%d", &r);
+    float r, area;
+    scanf("%f", &r);
     area = 3.14 * r**2;
-    printf("Area: %d\n", area);
+    printf("Area: %f\n", area);
     return 0;
 }
